Unify checkpoint mixin and parsable
