+++ parsonaut/__init__.py
@@ -1,4 +1,4 @@
-from .checkpoint_mixin import CheckpointMixin  # noqa: F401
-from .lazy import Choices, Lazy, Parsable  # noqa: F401
+from .lazy import Choices, Lazy  # noqa: F401
+from .parsable import Parsable  # noqa: F401
 from .parse import ArgumentParser  # noqa: F401
 from .serialization import Serializable  # noqa: F401
