+++ parsonaut/lazy.py
@@ -128,7 +128,6 @@ class Lazy(Generic[T, P], Serializable):
                 # fill in missing default
                 if value is Missing and (subtyp := get_args(typ)):
                     subtyp, *_ = subtyp
-                    assert issubclass(subtyp, Parsable)
                     value = Lazy.from_class(subtyp)  # type: ignore
                 # or ensure correct type
                 else:
@@ -253,74 +252,6 @@ class Lazy(Generic[T, P], Serializable):
             *args,
             **kwargs,
         )
-
-
-class ParsableMeta(type):
-    def __call__(cls, *args, **kwargs):
-
-        cfg = Lazy.from_class(cls, *args, skip_non_parsable=True, **kwargs)
-
-        # https://stackoverflow.com/a/73923070/8378586
-        obj = cls.__new__(cls, *args, **kwargs)
-        obj._cfg = cfg
-        # Initialize the final object
-        obj.__init__(*args, **kwargs)
-        return obj
-
-
-class class_or_instance_method(object):
-    def __init__(self, f):
-        self.f = f
-
-    def __get__(self, instance, owner):
-        if instance is not None:
-            class_or_instance = instance
-        else:
-            class_or_instance = owner
-
-        def newfunc(*args, **kwargs):
-            return self.f(class_or_instance, *args, **kwargs)
-
-        return newfunc
-
-
-class Parsable(Serializable, metaclass=ParsableMeta):
-
-    _cfg: Lazy
-
-    @class_or_instance_method
-    def as_lazy(cls_or_self, *args, **kwargs) -> Lazy:
-        if isinstance(cls_or_self, Parsable):
-            assert not args, "Cannot override once configured Parsable."
-            assert not kwargs, "Cannot override once configured Parsable."
-            return cls_or_self._cfg
-        else:
-            return Lazy.from_class(cls_or_self, *args, **kwargs)
-
-    def to_dict(
-        self,
-        with_class_tag: bool = False,
-        with_class_tag_as_str: bool = False,
-        flatten: bool = False,
-    ):
-        return self._cfg.to_dict(
-            with_class_tag=with_class_tag,
-            with_class_tag_as_str=with_class_tag_as_str,
-            flatten=flatten,
-        )
-
-    @classmethod
-    def from_dict(cls, dct) -> Lazy:
-        return Lazy.from_class(cls).from_dict(dct)
-
-    @classmethod
-    def parse_args(cls, *args, **kwargs) -> Lazy:
-        from .parse import ArgumentParser
-
-        parser = ArgumentParser()
-        parser.add_options(cls.as_lazy(*args, **kwargs))
-        params = parser.parse_args()
-        return params
 
 
 class Choices(Lazy, Enum):
